# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Knife Shop",
        "description": "Документация API для управления ножами",
        "version": "1.0.0"
    },
    "paths": {
        "/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Получить панель администратора",
                "operationId": "7cfa73c7f4bc906c66043dac50a3d549",
                "responses": {
                    "200": {
                        "description": "Страница админа"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/profile/edit": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Просмотр формы редактирования профиля",
                "operationId": "6fd54b3eb086b5a791127a54ab9cfa03",
                "responses": {
                    "200": {
                        "description": "Страница редактирования профиля"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Получить профиль пользователя",
                "operationId": "21d26f16795aefad77fc3cf09e8a7c65",
                "responses": {
                    "200": {
                        "description": "Форма редактирования профиля"
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Обновить профиль пользователя",
                "operationId": "b927bf00923c712a27589c6a510ebbf5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Профиль обновлен"
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Удалить аккаунт пользователя",
                "operationId": "3c62733109c5499344671a2ac26a1507",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Аккаунт удален"
                    },
                    "401": {
                        "description": "Неавторизован"
                    },
                    "403": {
                        "description": "Неверный пароль"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}